Move assert_vector_eq! into its own file
